requirements.txt
strands-agents
boto3
python-dotenv
strands-agents-tools
nest_asyncio
bedrock-agentcore
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
streamlit